Refactor import paths in workflow and settings pages
